Update VMVersionUpdater
Added If Else statement to check whether the server is running or not, and run the updated commands accordingly. This script will now only affect Running or Off servers and will leave servers in other states at their current version. This was added to prevent servers in the Off state from turning on unintentionally.